Reorganized classes
