+++ src/calculations/FirstMarketOrderCalculations.java
@@ -1,4 +1,4 @@
-package support;
+package calculations;
 
 public interface FirstMarketOrderCalculations {
