+++ src/calculations/ScalingTradeCalculations.java
@@ -1,4 +1,7 @@
-package support;
+package calculations;
+
+import data.TradeData;
+import data.TradeType;
 
 public class ScalingTradeCalculations extends TradeCalculations {
