+++ src/data/TradeDataInt.java
@@ -1,4 +1,4 @@
-package support;
+package data;
 
 public interface TradeDataInt {
